requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0